requirements.txt
gradio>=4.0.0
requests>=2.31.0
yfinance>=0.2.28
pandas>=2.0.0
python-dotenv>=1.0.0
